Revert "Last update"
This reverts commit 288964713b73c94d07f4def25580ca37c70a17b7, reversing
changes made to 641a6a89bd0662442147d09de54eaeb164594016.

+++ STAGGI-Budget-API/Controllers/BudgetController.cs
@@ -40,7 +40,7 @@ namespace STAGGI_Budget_API.Controllers
         }
 
         [HttpGet("{id}")]
-        public IActionResult GetById(int id)
+        public IActionResult GetById(long id)
         {
             string authorizationHeader = HttpContext.Request.Headers["Authorization"].FirstOrDefault()?.Substring(7);
             string userEmail = _authService.ValidateToken(authorizationHeader);
